% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{accessFit}
\alias{accessFit}
\title{accessFit constructor}
\usage{
accessFit(fit, I0, curve_dat, session_dat, speaker_word)
}
\arguments{
\item{fit}{\emph{nls.}
Logistic curve fit returned from
\code{\link[minpack.lm]{nlsLM}}.}

\item{I0}{\emph{numeric.}
Asymptotic intelligibility of the system under test.}

\item{curve_dat}{\emph{data.frame.}
Data frame of the treated data sent to the curve fitting
function.}

\item{session_dat}{\emph{list.}
List of raw data loaded from the session csv files.}

\item{speaker_word}{\emph{character vector.}
Vector of speaker word pairs for data used in accessFit}
}
\value{
\emph{accessFit.}
              Access fit object is a list based object containing the
              elements input to the constructor. Has following structure:
              \describe{
                  \item{fit}{Fit object from \code{\link[minpack.lm]{nlsLM}}. Characterizes
                  the intelligibility and access delay curves for the system under test.}
                  \item{I0}{Asymptotic intelligibility for the system under test.}
                  \item{curve_dat}{
                      A data frame with the following 3 variables: \describe{
                          \item{t}{PTT time, in seconds. Scaled such that 0 corresponds
                          with the start of the word under test and positive time
                          corresponds to the amount of time between PTT being pressed and
                          the word being spoken.}
                          \item{I}{Intelligibility}
                          \item{speaker}{Speaker and word combination, e.g. 'F1 hook'}}
                  }
                  \item{session_dat}{
                      A list of data frames with the following 8 variables: \describe{
                          \item{PTT_time}{PTT time in audio clip time, in seconds.
                          Corrected for audio interface latency.}
                          \item{PTT_start}{PTT time in audio clip time, in seconds.
                          Uncorrected for audio interface latency.}
                          \item{ptt_st_dly}{PTT time interval, in seconds.}
                          \item{P1_Int}{Intelligibility of P1.}
                          \item{P2_Int}{Intelligibility of P2.}
                          \item{m2e_latency}{Mouth-to-ear latency estimate for the given trial.}
                          \item{underRun}{Number of detected buffer under runs}
                          \item{overRun}{Number of detected buffer over runs}
                          }
                  }
                  \item{speaker_word}{Array of the speaker word combos used for the access
                  delay measurement.}
    }
}
\description{
Construct an access fit with model \code{fit}, asymptotic intelligibility
\code{I0}, curve fitting data \code{curve_dat}, and session data
\code{session_dat}.
Returns an 'accessFit' object. Compatible with generic functions:
\code{\link[base]{print}},
\code{\link[stats]{coef}},
\code{\link[stats]{confint}}.
}
\seealso{
\code{\link[minpack.lm]{nlsLM}}
}
