% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{aggregate_nls}
\alias{aggregate_nls}
\title{Aggregated access Fit constructor}
\usage{
aggregate_nls(sfits, pfits, sut_I0s)
}
\arguments{
\item{sfits}{\emph{nls.}
Logistic curve fit returned from
\code{\link[minpack.lm]{nlsLM}}.}

\item{pfits}{\emph{numeric.}
Data frame of the treated data sent to the curve fitting
function.}

\item{sut_I0s}{\emph{data.frame.}
Asymptotic intelligibility of the system under test.}
}
\value{
\emph{aggregate_nls.}
              Aggregate nls object provides aggregated fit parameters.
              Has following structure:
              \describe{
                  \item{call}{Fit formula.}
                  \item{lambda}{Linear combination of constituant lambda parameters.}
                  \item{t0}{Linear combination of constituant t0 parameters.}
                  \item{covar}{Propogated covariance matrix}
                  \item{words}{List of words in constituant fits.}
                  \item{orig_params}{List of SUT and PTT fit parameters.}
                  \item{SUT_I0s}{Asymptotic intelligibilities associated with SUT fits.}
    }
}
\description{
Construct an aggregated curve fit given SUT fits \code{sfits}, PTT fits
\code{pfits}, and the asymptotic intelligibilities of the SUT fits \code{sut_I0s}
}
\details{
Returns an 'aggregate_nls' object compatible with
\code{\link[base]{print}},
\code{\link[stats]{coef}},
\code{\link[stats]{confint}},
\code{\link[stats]{vcov}}
}
