% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analog_direct}
\alias{analog_direct}
\title{accessFit Object for analog direct LMR measurements}
\format{
See \code{\link[accessTime]{accessFit}} documentation
}
\usage{
analog_direct
}
\description{
Example access delay measurement data for a sample analog direct LMR system.
Corrected fit using ptt gate for the start of word correction.
}
\keyword{datasets}
