% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{apply_intell_theme}
\alias{apply_intell_theme}
\title{Apply intelligibility plot theme style to a ggplot object.}
\usage{
apply_intell_theme(p, title, xlim = c(-0.6, 2.5))
}
\arguments{
\item{p}{\emph{ggplot.} Plot object to apply theme to.}

\item{title}{\emph{character.} Title of plot.}

\item{xlim}{\emph{numeric.} Two element array defining x-axis limits.}
}
\description{
Apply intelligibility plot theme style to a ggplot object.
}
