% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_animations.R
\name{Audio_Plot}
\alias{Audio_Plot}
\title{Plot recording audio}
\usage{
Audio_Plot(
  rec,
  cp = NULL,
  PTT_time = NULL,
  title = "Audio",
  alpha = 1,
  time_limit = NULL,
  color = "black"
)
}
\arguments{
\item{rec}{\emph{Wave} \code{\link[tuneR]{Wave-class}} object containing
recording data.}

\item{cp}{\emph{data.frame} Data frame containing cutpoints for rec}

\item{title}{\emph{character} Title for audio plot}

\item{alpha}{\emph{numeric scalar} Value between 0 and 1 for opacity of
audio}

\item{time_limit}{\emph{numeric vector} Two element vector containing
limits for x-axis. See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{color}{\emph{character} Color for audio plot.}

\item{PTT_Time}{\emph{numeric scalar} Value at which vertical, black,
dotted line will be plotted to designate PTT being pressed.}
}
\description{
Plot audio data in \code{rec}. If cutpoints passed in via \code{cp}, then
P1 will be identified by a gray rectangle around it.
}
