% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{eval_access}
\alias{eval_access}
\title{Evaluate access times of accessFit object}
\usage{
eval_access(accFit, alpha, sys_dly_unc = 0.07/1.96 * 0.001)
}
\arguments{
\item{accFit}{\emph{accessFit.}
\code{\link[accessTime]{accessFit}} object}

\item{alpha}{\emph{numeric vector.}
Values between 0 and 1 that describe the fractional value of
the asymptotic intelligibility for an access delay
measurement.}

\item{sys_dly_unc}{Uncertainty of the measurement system delay. Accounts
for previous correction of PTT timing. Defaults to 0.07/2*1e-3.
This is a lower bound}
}
\value{
\code{Values} \emph{data.frame.}
              Data frame with three varaibles: alpha, access_time, and
              uncertainty. Each observation describes the access delay
              associated with an intelligibility of \code{alpha}*\code{I0}.
              Note that access delay values are not dependent on I0.
}
\description{
Evaluates an access delay function derived from an access fit object,
\code{accFit}, at the values \code{alpha}.
}
\examples{
# Evaluation of data collected for NISTIR XXXX

# Define alpha vector
alphas <- seq(from=0.01, to=0.99, by=0.01)

# Define system delay uncertainty
sys_dly_unc <- 1e-3*0.07/1.96

access_delays <- eval_access(analog_direct, alphas, sys_dly_unc)

}
\references{
Pieper J, Frey J, Greene C, Soetan Z, Thompson T, Bradshaw D,
Voran S (2019) Mission Critical Voice QoE Access Time Measurement Methods.
\emph{NISTIR}. URL \url{https://doi.org/10.6028/NIST.IR.XXXX}
}
\seealso{
\code{\link[accessTime]{accessFit}}
}
