% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{eval_intell}
\alias{eval_intell}
\title{Evaluate an intelligibility of accessFit object}
\usage{
eval_intell(accFit, t)
}
\arguments{
\item{accFit}{\emph{accessFit.}
\code{\link[accessTime]{accessFit}} object.}

\item{t}{\emph{numeric vector.}
Times at which to evaluate the intelligibility curve}
}
\value{
\code{Values} \emph{data.frame.}
              Data frame with two variables: \code{t} and \code{I}. Each
              observation describes the intelligibility of the system under
              test at time \code{t}.
}
\description{
Evaluates an intelligibility function, I(t), defined by the accessFit
object, accFit, at the times defined by t.
}
\examples{
#Evaluation of data collected for NISTIR XXXX

# Define time vector
times <- seq(from = -0.5, to = 2, by = 0.01)

intelligibility_values <- eval_intell(ptt_gate,times)

}
\references{
Pieper J, Frey J, Greene C, Soetan Z, Thompson T, Bradshaw D,
Voran S (2019) Mission Critical Voice QoE Access Time Measurement Methods.
\emph{NISTIR}. URL \url{https://doi.org/10.6028/NIST.IR.XXXX}
}
\seealso{
\code{\link[accessTime]{accessFit}}
}
