% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_animations.R
\name{find_example_rec}
\alias{find_example_rec}
\title{Find example recordings in access time data}
\usage{
find_example_rec(session_data, ptt_time, Ival, ptt_step = 0.02)
}
\arguments{
\item{session_data}{\emph{data.frame} Data frame for access time test}

\item{ptt_time}{\emph{numeric scalar} PTT time to search around}

\item{Ival}{\emph{numeric scalar} Intelligibility value (0<\code{Ival}<1)
that example P1 should be near}

\item{ptt_step}{\emph{numeric scalar} PTT time increment for given test.}
}
\description{
Search \code{session_data} for an example recording at a given PTT time
that has P1 intelligibility close to \code{Ival}.
}
