% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{fit_accessData}
\alias{fit_accessData}
\title{Treat and fit access delay data to a logistic curve}
\usage{
fit_accessData(dat, cutPoints, speaker_word, fs)
}
\arguments{
\item{dat}{\emph{list.}
List of data frames containing the access delay data for a
test. Data frame must have same variables as the csv files
output from an access delay measurement.}

\item{cutPoints}{\emph{list.}
List of data frames containing the cutpoints of the audio clips
used in testing. Cutpoints describe the samples at which P1
and P2 start and end in the audio clip.}

\item{speaker_word}{\emph{character array.}
Vector of speaKer and word combination for the words under test.
Example: 'F1 hook'}

\item{fs}{\emph{numeric.}
Sampling rate of the audio clips used in testing (Hz).
Allows sample number to be converted to a time in seconds.}
}
\value{
\emph{accessFit.} object, a list containing the following
              elements:

\code{fit} \emph{nls.}
              Logistic curve fit object returned from
              \code{\link[minpack.lm]{nlsLM}}.

\code{I0} \emph{numeric.}
              Asymptotic intelligibility of the system under test.

\code{curve_dat} \emph{data.frame.}
              Data frame of the treated data sent to the curve fitting
              function.

\code{session_dat} \emph{list.}
              List of raw data loaded from the session csv files. Same as
              \code{dat}.

\code{speaker_word} \emph{character array.}
              Array of the speaker and word combination (e.g. 'F1 hook')
              for each audio clip.
}
\description{
Treat and fit access delay data to a logistic curve. First, recenters timing
of access delay data so that the start of \code{P1} is at 0 seconds.
Then, calculates the asymptotic intelligiblity of each word and fits a single
logistic function across all of the access delay data. Returns an
'accessFit' object. Compatible with generic functions:
\code{\link[base]{print}}, \code{\link[stats]{coef}},
\code{\link[stats]{confint}}. See also
\code{\link[accessTime]{read_accessData}} to read access delay data to be fit,
and also \code{\link[accessTime]{process_accessData}} which is a wrapper
function for both \code{read_accessData} and \code{fit_accessData}.
}
\examples{
# Path to included raw data
raw_path <- system.file("extdata", "Measurements", package = "accessTime")

# All included raw data tests
raw_testNames <- list.files(raw_path)

# Full path to each test
raw_testPath <- file.path(raw_path,raw_testNames)

# Session CSV Files for each raw test
raw_sessionFiles <- lapply(raw_testPath,list.files)

# Full path to all session files for each test
sessionFiles <- mapply(function(x,y){file.path(x,y)},raw_testPath,raw_sessionFiles,SIMPLIFY = FALSE)
names(sessionFiles) <- raw_testNames

# Directory to audio clip wav files and cut points
cutDir <- system.file("extdata", "Audio_Clips", package="accessTime")

# Read session data
session_dat <- lapply(sessionFiles,read_accessData,cutDir)

# Fit session data
fit_data <- lapply(session_dat,function(access_data){
                          fit_accessData(dat = access_data$dat,
                                         cutPoints = access_data$cutPoints,
                                         speaker_word = access_data$speaker_word,
                                         fs = access_data$fs)
                          })

}
\seealso{
\code{\link[minpack.lm]{nlsLM}}
}
