% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lte}
\alias{lte}
\title{accessFit object for an in-development LTE system measurements}
\format{
See \code{\link[accessTime]{accessFit}} documentation
}
\usage{
lte
}
\description{
Example access delay measurement data for a sample LTE system.
Corrected fit using ptt gate for the start of word correction.
}
\keyword{datasets}
