% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_animations.R
\name{make_frame}
\alias{make_frame}
\title{Make access time animation frame}
\usage{
make_frame(
  dat_subset,
  x_axis = round(seq(-0.6, 2, by = 0.2), 1),
  example = FALSE,
  fit = NULL,
  alpha = 1
)
}
\arguments{
\item{dat_subset}{\emph{data.frame} Data frame containing access data to
  plot. Must contain columns named \code{t} and \code{I}, representing PTT
  time and intelligibility respecitively.
i}

\item{x_axis}{\emph{numeric vector} Breaks for x-axis (see
\code{\link[ggplot2]{scale_x_continuous}})}

\item{example}{\emph{Numeric} Row number of \code{dat_subset} for the
example recording}

\item{fit}{\emph{data.frame} Data frame containing curve fit data. Must
have columns named \code{t} and \code{I} representing time and
intelligibility respectively.}

\item{alpha}{\emph{numeric.} Opacity of plot objects.}
}
\description{
Make single animation frame from data in \code{dat_subset}. If
\code{example} is passed in, the example point is marked by a red point. If
\code{fit} is passed in, it is plotted over the data points from
\code{dat_subset}.
}
