% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_animations.R
\name{make_test_frames}
\alias{make_test_frames}
\title{Create animations frames of access time tests.}
\usage{
make_test_frames(
  accFit,
  alpha_lvls = c(0.01, 0.33, 0.66, 0.99),
  frame_dir = "frames",
  Audio_Data = NULL
)
}
\arguments{
\item{accFit}{\emph{accessFit} \code{\link[accessTime]{accessFit}} object.}

\item{alpha_lvls}{\emph{Numeric Vector} Fractional intelligibility levels to
show example recordings. Values must be in range (0,1). Defaults to
\code{alpha_lvls = c(0.01, 0.33, 0.66, 0.99)}.}

\item{frame_dir}{\emph{character} Path where animation frames will be
stored. Defaults to \code{frame_dir = "frames"}.}

\item{Audio_Data}{\emph{list} List with following structure: \describe{
\item{rec}{\emph{Wave} \code{\link[tuneR]{Wave-class}} object}
\item{cp}{\emph{data.frame} Data frame containing cutpoints for rec}
\item{rx_path}{\emph{character} Path where received audio files stored} }}
}
\description{
Create an animation frames using data from \code{accFit}. Iterate through
PTT times, and incrementally plot intelligibility scores. At each element of
\code{alpha_lvls} specific trials will be identified that have
intelligibilites close to the alpha_lvl on the intelligibility curve. If
\code{Audio_Data} is passed as an argument, example recordings will also be
plotted and played during the animation. Note that the animation plots data
for only one talker. Note that the accFit must be data for a single
word/talker pair. Animations will not generate for accessFit objects with
multiple talkers/words.
}
\seealso{
\code{\link[accessTime]{make_access_animation}}
}
