% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p25_direct}
\alias{p25_direct}
\title{accessFit Object for P25 direct LMR measurements}
\format{\code{\link[accessTime]{accessFit}} objects have the following
components:

\describe{
    \item{fit}{Fit object from \code{\link[minpack.lm]{nlsLM}}. Characterizes
    the intelligibility and access delay curves for the system under test.}
    \item{I0}{Asymptotic intelligibility for the system under test.}
    \item{curve_dat}{
        A data frame with the following 3 variables: \describe{
            \item{t}{PTT time, in seconds. Scaled such that 0 corresponds
            with the start of the word under test and positive time
            corresponds to the amount of time between PTT being pressed and
            the word being spoken.}
            \item{I}{Intelligibility}
            \item{speaker}{Speaker and word combination, e.g. 'F1 hook'}}
    }
    \item{session_dat}{
        A list of data frames with the following 8 variables: \describe{
            \item{PTT_time}{PTT time in audio clip time, in seconds.
            Corrected for audio interface latency.}
            \item{PTT_start}{PTT time in audio clip time, in seconds.
            Uncorrected for audio interface latency.}
            \item{ptt_st_dly}{PTT time interval, in seconds.}
            \item{P1_Int}{Intelligibility of P1.}
            \item{P2_Int}{Intelligibility of P2.}
            \item{m2e_latency}{Mouth-to-ear latency estimate for the given trial.}
            \item{underRun}{Number of detected buffer under runs}
            \item{overRun}{Number of detected buffer over runs}
            }
    }
    \item{speaker_word}{Array of the speaker word combos used for the access
    delay measurement.}
}}
\usage{
p25_direct
}
\description{
Example access delay measurement data for a sample P25 direct LMR system.
}
\keyword{datasets}
