% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p25_trunked_phase2}
\alias{p25_trunked_phase2}
\title{accessFit Object for P25 Trunked Phase 2 LMR measurements}
\format{
See \code{\link[accessTime]{accessFit}} documentation
}
\usage{
p25_trunked_phase2
}
\description{
Example access delay measurement data for a sample P25 Trunked Phase 2 LMR system.
Corrected fit using ptt gate for the start of word correction.
}
\keyword{datasets}
