% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_access}
\alias{plot_access}
\title{Plot access delay}
\usage{
plot_access(
  ...,
  plotlist = NULL,
  curve_names = NULL,
  title = "Access Delay Curves",
  xlim = c(0.5, 0.99),
  ylim = NULL,
  intell = FALSE,
  level = 0.95
)
}
\arguments{
\item{...}{list of accessFit objects to plot.}

\item{plotlist}{\emph{list.}
List of accessFit objects to plot access delay data for.}

\item{curve_names}{\emph{character.}
Array of names for each curve. Used to label curves in plot legend.}

\item{title}{\emph{character.}
Plot title.}

\item{xlim}{\emph{numeric.}
X-axis limits (alpha values between 0 and 1) for plot.}

\item{ylim}{\emph{numeric.}
Y-axis limits (time in seconds).}

\item{intell}{\emph{logical.} Determines if x-axis is intelligibility or
alpha level.}

\item{level}{\emph{numeric.} Confidence level for confidence intervals
plotted on access delay curves.Value between 0 and 1.}
}
\value{
\code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot access delay for \emph{accessFit} objects. Can plot multiple access
delay curves at once.
}
\examples{
# Plot access for analog direct data
plot_access(analog_direct)

# Plot access for P25 Phase 1 and Phase 2 Trunked, using intelligibility instead of alpha for x-axis
plot_access(p25_trunked_phase1,p25_trunked_phase2,curve_names = c("Phase 1", "Phase 2"),intell = TRUE)
}
