% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{plot_accessCurve}
\alias{plot_accessCurve}
\title{Plot access delay curve}
\usage{
plot_accessCurve(access_dat, title = "")
}
\arguments{
\item{access_dat}{\emph{data.frame}
Data frame with the following three variables:
\describe{
\item{alpha}{Value between 0 and 1. Describes a fraction of
the asymptotic intelligibility, I0, achieved.}
\item{access_time}{Access delay values, in seconds.}
\item{uncertainty}{Standard, unexpanded, uncertainty of access
delay values, in seconds.}}}

\item{title}{\emph{character.}
Title of plot}
}
\value{
\emph{ggplot.}
              ggplot object of plot.

 @import ggplot2
}
\description{
Plot access delay curve values calculated from an accessFit object.
}
\examples{

# Define alpha vector
alphas <- seq(from = 0.5, to = 0.99, by = 0.01)

# Define system delay uncertainty
sys_dly_unc <- 1e-3*0.07/1.96

# Evaluate access delays
access_delays <- eval_access(ptt_gate,alphas,sys_dly_unc)

plot_accessCurve(access_delays,"PTT Gate Access Delay")

}
