% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_intelligibility}
\alias{plot_intelligibility}
\title{Plot intelligibility results}
\usage{
plot_intelligibility(
  accFit,
  xlim = c(-0.6, 2.5),
  word_curves = FALSE,
  title = "Intelligibility Curve",
  include_raw = FALSE
)
}
\arguments{
\item{accFit}{\emph{accessFit.}
Access fit object.}

\item{xlim}{\emph{numeric.}
X-axis limits (time in seconds) for plot.}

\item{word_curves}{\emph{logical.}
Determines if word curves are plotted or not. Throws warning if
\emph{accessFit} does not have valid data for word curves
(e.g. LEG, SUT, or PTT fit).}

\item{title}{\emph{character.}
Title of the plot}

\item{include_raw}{\emph{logical.}
Plot all raw intelligibility data contributing to intelligibility curve.}
}
\value{
\code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot intelligibility results for a given access fit object. Plots the
intelligibility curve, with options to plot individual words contributing
to the final fit as well as the raw intelligibility data.
}
\examples{
# Plot intelligibility curve for LTE data
plot_intelligibility(lte)
# Plot intelligibility curve for LTE data with individual word curves
plot_intelligibility(lte,word_curves=TRUE)

# Plot intelligibility curve analog direct data with raw intelligibility points
plot_intelligibility(analog_direct,include_raw = TRUE)
}
