% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_raw_intells}
\alias{plot_raw_intells}
\title{Plot raw intelligibility data}
\usage{
plot_raw_intells(
  accFit,
  xlim = c(-0.6, 2.5),
  title = "Raw Intelligibilities for Select Words"
)
}
\arguments{
\item{accFit}{\emph{accessFit.}
Access fit object.}

\item{xlim}{\emph{numeric.}
X-axis limits (time in seconds) for plot.}

\item{title}{\emph{character.}
Title of the plot}
}
\value{
\code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot raw intelligibility data from an \emph{accessFit} object. For PTT or SUT
type fit plots intelligibilities. For COR creates plots for PTT and SUT
separately. For LEG plots SUT raw data.
}
\examples{
plot_raw_intells(analog_direct)
}
