% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{plot_techIntCurve}
\alias{plot_techIntCurve}
\title{Plot intelligibility curve for a technology}
\usage{
plot_techIntCurve(accFit, int_curve, title = "")
}
\arguments{
\item{accFit}{\emph{accessFit.}
\code{\link[accessTime]{accessFit}} object.}

\item{int_curve}{\emph{data.frame.}
Data frame of intelligibility values from
\code{\link[accessTime]{eval_intell}} function. Must contain two variables:
\describe{
    \item{t}{PTT time, in seconds. Time corrected for audio
    clips such that 0 seconds corresponds to PTT being
    pressed right at the beginning of the word being spoken.}
    \item{I}{Intelligibility}
 }}

\item{title}{\emph{character.} Title of the plot}
}
\value{
\emph{ggplot}
}
\description{
Plot intelligibility curve made up of multiple words across one technology.
}
\examples{
# Times to evaluate intelligibility at
times <- seq(from = -0.5, to = 2, by = 0.01)

# Intelligibility as function of time
intelligibility_values <- eval_intell(ptt_gate,times)


plot_techIntCurve(ptt_gate,intelligibility_values,
                  title="PTT Gate Intelligibility Curve")
}
