% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{process_session_data}
\alias{process_session_data}
\title{Process individual accessFits for each match to id_strs.}
\usage{
process_session_data(id_strs, test_type, dat_path = "data", maxiter = 50)
}
\arguments{
\item{id_strs}{\emph{character array.}
Identifying character strings for sessions to process.}

\item{test_type}{\emph{character.}
Type of access fit to return. Must be one of PTT, SUT, or
LEG. See \code{\link[accessTime]{accessFit}}.}

\item{dat_path}{\emph{character array.}
Parent directory containg csv to use.}

\item{maxiter}{\emph{numeric.}
Maximum number of iterations passed to fitting algorithm. Defaults to 50}
}
\value{
List of access fit objects matching \code{id_strs}.
}
\description{
Process individual accessFits for each match to id_strs. Each fit is
classified as test_type. This function is designed to make creating
corrected technology fits easier.
}
\examples{
# get PTT Gate fits

 dat_path <- system.file("extdata", package="accessTime")
 ptt_fits <- process_session_data(
   "capture_PTT-gate_14-May-2021_07-30-20",
   test_type = "PTT",
   dat_path = dat_path
   )

 # Get SUT fits
 sut_fits <- process_session_data(
   "capture_LTE_14-Apr-2021_16-11-22",
   test_type="SUT",
   dat_path = dat_path
 )

 # Get technology fit
 lte <- techFit(
   sut_fits = sut_fits,
   ptt_fits = ptt_fits
 )

}
