% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptt_gate}
\alias{ptt_gate}
\title{List of accessFit objects for PTT Gate measurements}
\format{
See \code{\link[accessTime]{accessFit}} documentation
}
\usage{
ptt_gate
}
\description{
Example access delay measurement data for a simple radio replacement,
known as a PTT gate. Used for start of word correction via
\code{\link[accessTime]{techFit}}.
}
\details{
This data can be easily loaded from raw data using
\code{\link[accessTime]{process_session_data}} with
\code{id_strs = "capture_PTT-gate_14-May-2021_07-30-20"}, see example.
}
\examples{
dat_path <- system.file("extdata", package="accessTime")
# Load the ptt gate data
ptt_gate <- process_session_data("capture_PTT-gate_14-May-2021_07-30-20","PTT",dat_path=dat_path)

# Load data for SUT (analog direct mode here)
analog_direct_wut <- process_session_data("capture_Analog_12-Nov-2020_08-26-11", "SUT", dat_path=dat_path)
# Use the ptt gate data to generate a corrected technology fit
analog_direct <- techFit(analog_direct_wut,ptt_gate)
}
\keyword{datasets}
