% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{read_accessData}
\alias{read_accessData}
\title{Read access delay csv files and associated audio cut point files}
\usage{
read_accessData(sessionFiles, cutDir)
}
\arguments{
\item{sessionFiles}{\emph{character array.}
Array of csv file names associated with a particular access
delay test.}

\item{cutDir}{\emph{character.}
Path of the folder that contains the cutpoint files for the
audio clips used in the access delay test.}
}
\value{
\emph{List} with the following elements:

\code{dat} \emph{list.}
              List of data frames containing the raw access delay data of
              each csv file in \code{sessionFiles}.

\code{cutPoints} \emph{list.}
              List of data frames containing the cutpoints of the audio clips
              used in testing. Cutpoints describe the samples at which P1
              and P2 start and end in the audio clip.

\code{fs} \emph{numeric.}
              Sampling rate of the audio clips used in testing (Hz).

\code{audio_clips} \emph{character array.}
              Array of audio clips associated with each csv file in
              \code{sessionFiles}.

\code{speaker_word} \emph{character array.}
              Array of the speaker and word combination (e.g. 'F1 hook')
              for each audio clip.
}
\description{
Read the access delay csv files in \code{sessionFiles}, identify the audio
              clips used for the tests, and load in the associated cut
              point files from \code{cutDir}.
}
\examples{
# Path to included raw data
raw_path <- system.file("extdata", "Measurements", package = "accessTime")

# All included raw data tests
raw_testNames <- list.files(raw_path)

# Full path to each test
raw_testPath <- file.path(raw_path,raw_testNames)

# Session CSV Files for each raw test
raw_sessionFiles <- lapply(raw_testPath,list.files)

# Full path to all session files for each test
sessionFiles <- mapply(function(x,y){file.path(x,y)},raw_testPath,raw_sessionFiles,SIMPLIFY = FALSE)
names(sessionFiles) <- raw_testNames

# Directory to audio clip wav files and cut points
cutDir <- system.file("extdata", "Audio_Clips", package="accessTime")

# Read session data
session_dat <- lapply(sessionFiles,read_accessData,cutDir)

# Fit session data
fit_data <- lapply(session_dat,function(access_data){
                          fit_accessData(dat = access_data$dat,
                                         cutPoints = access_data$cutPoints,
                                         speaker_word = access_data$speaker_word,
                                         fs = access_data$fs)
                          })

}
