% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{sortSpeakers}
\alias{sortSpeakers}
\title{Sort session file names or audio clip names alphabetically by speaker,
word pairs.}
\usage{
sortSpeakers(sf, ext = ".csv")
}
\arguments{
\item{sf}{\emph{character vector.}
Vector of strings of form '*_Speaker_bXX_wXX_word*'.
Examples:
\code{"capture_Analog_Direct_24-Apr-2019_14-39-14_F1_b39_w4_hook.csv"}
\code{"capture_Analog_Direct_23-Apr-2019_15-32-52_F3_b15_w5_west.csv"}
\code{"capture_Analog_Direct_23-Apr-2019_11-54-00_M3_b22_w3_cop.csv"}
\code{"capture_Analog_Direct_23-Apr-2019_08-12-08_M4_b18_w4_pay.csv"}}

\item{ext}{\emph{character.}
File extension of the elements of \code{sf}.}
}
\value{
\emph{List} with following elements:

\code{ordered} \emph{character vector.}
              Alphabetized version of \code{sf}.

\code{speaker_word} \emph{character vector.}
              The speaker word combination extracted from \code{sf}.

\code{order} \emph{numeric vector.}
              Indices such that \code{ordered = sf[order]}.
}
\description{
Sort file names that end in format of '_Speaker_bXX_wXX_word.'
alphabetically by speaker, word.
}
