% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_animations.R
\name{stitch_frames}
\alias{stitch_frames}
\title{Stitch animation frames into a video file}
\usage{
stitch_frames(
  frame_dir,
  output,
  frame_rate = 10,
  curve_fit_audio = "",
  video_codec = "libx264",
  audio_codec = "aac",
  tx_examples = NULL,
  powerpoint = FALSE,
  extra_settings = NULL,
  output_directory = ""
)
}
\arguments{
\item{frame_dir}{\emph{character} Path where animation frames are stored.}

\item{output}{\emph{character} Name of output video file.}

\item{frame_rate}{\emph{numeric} Frame rate of input frames in animation.
Final output video frame rate is fixed at 30 fps. This parameter controls
how long each frame is displayed for, but does not affect final video
frame rate.}

\item{curve_fit_audio}{\emph{character} Path to audio to play when curve fit
image is displayed.}

\item{video_codec}{\emph{character} Video codec to use for output video}

\item{audio_codec}{\emph{character} Audio codec to use for output video}

\item{tx_examples}{\emph{numeric array.} Array containing which examples
should play but Tx and Rx audio together to demonstrate mouth-to-ear
latency. Default is \code{NULL}, which causes the no examples to have both
audio files.}

\item{powerpoint}{\emph{logical.} Indicator to optimize for powerpoint
presentations. Will override codec settings add some options to make the
video import into power point presentations well.}

\item{extra_settings}{\emph{character.} Pass arbitrary options to ffmpeg}

\item{output_directory}{\emph{character.} Directory to save intermediate
files to}
}
\description{
Creates an animation from frames in \code{frame_dir} and outputs video file
as \code{output}.
}
\details{

}
\section{frame_dir}{
 The directory where the frames to put in the animation
 are stored. The animation contains any number of sections where each
 section contains any number of frames.

 Each section has a main part and an, optional, example part. The frames in
 the main part are named SectionSS_FrameFFF.png where SS is the zero padded
 section number and FFF is the zero padded frame number. The example part
 contains a single frame named SectionSS_Example.png where SS is the
 section number. The example part also contains an audio clip named
 receive_exampleSS.wav where SS is again the section number. If either the
 example frame or audio is missing there is no example part.

 For the end of the clip a special section is created with the frame
 curve_fit.png which is shown for a duration of 1/2 a second.

 Once all sections are completed they are concatinated together to create
 the clip.
}

\section{FFMPEG and Python}{
 In order for the animation frames to be stitched
 into a single video, both FFMPEG and Python are required to be in the
 system path. If they are not in the system path, an error will be thrown.
}

\seealso{
\code{\link[accessTime]{make_access_animation}}
}
