% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{techFit}
\alias{techFit}
\title{Technology access Fit constructor}
\usage{
techFit(sut_fits, ptt_fits)
}
\arguments{
\item{sut_fits}{\emph{accessFit.}
List of accessFit objects representing the WUTs through a respective SUT.}

\item{ptt_fits}{\emph{accessFit.}
List of accessFit objects representing the WUTs through a PTT gate.}
}
\value{
\emph{accessFit.}
              Technology fit object is a list based object containing the
              elements input to the constructor. Has following structure:
              \describe{
                  \item{fit}{An aggregate_nls fit of the constituant fits.}
                  \item{I0}{Arithmetic mean of asymptotic intelligibilites of the SUT fits.}
                  \item{speaker_word}{List of keywords used in the SUT trial.}
                  \item{test_type}{Test type. Set to "COR". Using this constructor implies a corrected fit.}
    }
}
\description{
Construct a technology fit given SUT curve fits \code{sut_fits}, and PTT
curve fits \code{ptt_fits}
Returns an 'accessFit' object. Compatible with generic functions:
\code{\link[base]{print}},
\code{\link[stats]{coef}},
\code{\link[stats]{confint}},
\code{\link[stats]{vcov}}.
}
\examples{
dat_path <- system.file("extdata", package="accessTime")
 ptt_fits <- process_session_data(
   "capture_PTT-gate_14-May-2021_07-30-20",
   test_type = "PTT",
   dat_path = dat_path
   )

 # Get SUT fits
 sut_fits <- process_session_data(
   "capture_LTE_14-Apr-2021_16-11-22",
   test_type="SUT",
   dat_path = dat_path
 )

 # Get technology fit
 lte <- techFit(
   sut_fits = sut_fits,
   ptt_fits = ptt_fits
 )
}
