% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p25_direct}
\alias{p25_direct}
\title{accessFit Object for P25 direct LMR measurements}
\format{
See \code{\link[accessTime]{accessFit}} documentation
}
\usage{
p25_direct
}
\description{
Example access delay measurement data for a sample P25 direct LMR system.
Corrected fit using ptt gate for the start of word correction.
}
\keyword{datasets}
