% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{process_accessData}
\alias{process_accessData}
\title{Wrapper function to read, process, and fit access delay data}
\usage{
process_accessData(sessionFiles, cutDir)
}
\arguments{
\item{sessionFiles}{\emph{character array.}
Array of csv file names associated with a particular access
delay test.}

\item{cutDir}{\emph{character.}
Name of the directory where cutpoint files are stored for the
audio clips used in an access delay test.}
}
\value{
\emph{accessFit.}
              Access fit object is a list containing the following
              elements:

\code{fit} \emph{nls.}
              Logistic curve fit returned from
              \code{\link[minpack.lm]{nlsLM}}.

\code{I0} \emph{numeric.}
              Asymptotic intelligibility of the system under test.

\code{curve_dat} \emph{data.frame.}
              Data frame of the treated data sent to the curve fitting
              function.

\code{session_dat} \emph{list.}
              List of raw data loaded from the session csv files.
}
\description{
Read in access delay data, treat it, and fit a logistic function to the
treated data. Relies on calls to \code{\link[accessTime]{read_accessData}}
and \code{\link[accessTime]{fit_accessData}}. Returns an 'accessFit' object.
Compatible with generic functions: \code{\link[base]{print}},
\code{\link[stats]{coef}}, \code{\link[stats]{confint}}.
}
\examples{
# Path to included raw data
raw_path <- system.file("extdata", "Measurements", package = "accessTime")

# All included raw data tests
raw_testNames <- list.files(raw_path)

# Full path to each test
raw_testPath <- file.path(raw_path,raw_testNames)

# Session CSV Files for each raw test
raw_sessionFiles <- lapply(raw_testPath,list.files)

# Full path to all session files for each test
sessionFiles <- mapply(function(x,y){file.path(x,y)},raw_testPath,raw_sessionFiles,SIMPLIFY = FALSE)
names(sessionFiles) <- raw_testNames

# Directory to audio clip wav files and cut points
cutDir <- system.file("extdata", "Audio_Clips", package="accessTime")

# Read, process, and fit session data
fit_data <- lapply(sessionFiles,process_accessData,cutDir)

}
\seealso{
\code{\link[minpack.lm]{nlsLM}}
}
