% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{read_accessData}
\alias{read_accessData}
\title{Read access delay csv files and associated audio cut point files}
\usage{
read_accessData(sessionFiles, cutFiles)
}
\arguments{
\item{sessionFiles}{\emph{character array.} Array of csv file names
associated with a particular access delay test.}

\item{cutFiles}{\emph{character array.} Array of cutpoint file names.}
}
\value{
\emph{List} with the following elements:

\code{dat} \emph{list.} List of data frames containing the raw
 access delay data of each csv file in \code{sessionFiles}.

\code{cutPoints} \emph{list.} List of data frames containing the
 cutpoints of the audio clips used in testing. Cutpoints describe the
 samples at which P1 and P2 start and end in the audio clip.

\code{fs} \emph{numeric.} Sampling rate of the audio clips used in
 testing (Hz).

\code{audio_clips} \emph{character array.} Array of audio clips
 associated with each csv file in \code{sessionFiles}.

\code{speaker_word} \emph{character array.} Array of the speaker and
 word combination (e.g. 'F1 hook') for each audio clip.
}
\description{
Read the access delay csv files in \code{sessionFiles}, identify the audio
clips used for the tests, and load in the associated cut point files from
\code{cutDir}.
}
\examples{

# Path to included raw data
dat_path <- system.file("extdata", package="accessTime")

# Analog Direct csv files
data_files <- find_session_csv(
  "capture_Analog_12-Nov-2020_08-26-11",
  dat_path
  )

# Get the session and cut file for one WUT
sessionFile <- data_files$sessions[1]
cutFile <- data_files$cut[1]
# Read session data
sess_dat <- read_accessData(sessionFile, cutFile)

# Fit session data
fit_dat <- fit_accessData(
  dat=sess_dat$dat,
  test_type="SUT",
  cutPoints=sess_dat$cutPoints,
  speaker_word=sess_dat$speaker_word,
  fs=sess_dat$fs
  )

}
